requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Axolotl and related
git+https://github.com/OpenAccess-AI-Collective/axolotl.git
peft>=0.7.0
bitsandbytes>=0.41.0
flash-attn>=2.3.0
deepspeed>=0.12.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Utilities
pyyaml>=6.0
wandb>=0.15.0
tqdm>=4.66.0
requests>=2.31.0

# Optional optimizations
xformers>=0.0.22
optimum>=1.14.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 